Added support for `struct` types
